Heavily optimize vulnerability detection setting
Using Sets, the #include? lookup later on goes from being an O(n)
operation to being O(1).  Considering that this takes the program from
O(n * m) time [where n is the number of plugins and m is the number of
detections for each plugin] to O(n) time, this is a huge win.